+++ core/components/superboxselect/model/superboxselect/superboxselect.class.php
@@ -3,7 +3,7 @@
  * SuperBoxSelect classfile
  *
  * Copyright 2011-2016 by Benjamin Vauchel <contact@omycode.fr>
- * Copyright 2016-2020 by Thomas Jakobi <thomas.jakobi@partout.info>
+ * Copyright 2016-2021 by Thomas Jakobi <thomas.jakobi@partout.info>
  *
  * @package superboxselect
  * @subpackage classfile
@@ -30,7 +30,7 @@ class SuperBoxSelect
      * The version
      * @var string $version
      */
-    public $version = '2.4.1-pl2';
+    public $version = '2.4.2';
 
     /**
      * The class config