+++ core/components/superboxselect/docs/changelog.md
@@ -1,93 +1,134 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.4.2] - TBA
+
 ### Changed
+
 - Fix retrieving resrources from a different context deeper than one level [#24].
 - Prevent E_NOTICE errors from smarty.
 - Fix drag & drop issues [#25].
 
 ## [2.4.1] - 2020-08-26
+
 ### Changed
+
 - Fix not working boolean template variable options [#23].
 
 ## [2.4.0] - 2020-03-22
+
 ### Added
+
 - Allow other values to be saved in the SuperBoxSelect than the resource id
+
 ### Changed
+
 - Restore pageSize for multi selects disabled in SuperBoxSelect 2.3.8
 - Fix showing 'The changes you made may not be saved' alert, even when the SuperBoxSelect value was not changed.
 
 ## [2.3.8] - 2020-01-28
+
 ### Changed
+
 - Fix a limit/pageSize issue
 
 ## [2.3.7] - 2020-01-14
+
 ### Changed
+
 - Fix autocomplete is not working on user list
 
 ## [2.3.6] - 2020-01-04
+
 ### Changed
+
 - Fix sorting of user selector [#15]. Thanks to @hugopeek
 
 ## [2.3.5] - 2019-10-19
+
 ### Changed
+
 - Fix an issue that stack items option can't be disabled
 
 ## [2.3.4] - 2019-09-29
+
 ### Changed
+
 - Update Sortable.js to 1.10.0 to fix drag & drop ordering
 
 ## [2.3.3] - 2019-05-06
+
 ### Changed
+
 - Fix drag & drop ordering after deleting a list element
 
 ## [2.3.2] - 2019-04-27
+
 ### Changed
+
 - Fix an issue, when an external package does not exist
 - Fix drag & drop ordering
 
 ## [2.3.1] - 2019-03-20
+
 ### Changed
+
 - Fix input options for MODX Revolution 2.7+
 - Use less space in input options
 
 ## [2.3.0] - 2019-02-17
+
 ### Added
+
 - Enable stacking of the SuperBoxSelect values per template variable option
 
 ## [2.2.2] - 2017-02-03
+
 ### Changed
+
 - Fixing limit to related context option
 
 ## [2.2.1] - 2017-01-29
+
 ### Changed
+
 - Fixing disabling the pagination
 
 ## [2.2.0] - 2016-12-21
+
 ### Added
+
 - Sort the items by drag & drop
 
 ## [2.1.2] - 2016-10-13
+
 ### Changed
+
 - Fix backwards compatibility issue
 
 ## [2.1.1] - 2016-10-11
+
 ### Changed
+
 - Fix single value combo not saving the value
 
 ## [2.1.0] - 2016-09-08
+
 ### Added
+
 - The superboxselect is not restricted anymore to list only resources
 - Single value
 - Two default types (resources, users)
 - Add own package based types
 
 ## [2.0.0] - 2016-06-27
+
 ### Added
+
 - Input options for template variables
 - Pagination of the superboxselect
 - Maximum number of selectable resources
@@ -96,26 +137,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Strictly separate the extra folders from the MODX core
 
 ## [1.0.1-pl2] - 2012-06-16
+
 ### Changed
+
 - Fix fetch resource bug
 
 ## [1.0.1-pl1] - 2012-06-16
+
 ### Changed
+
 - Fix height component to display properly multiple rows
 
 ## [1.0.1-rc2] - 2012-02-01
+
 ### Changed
+
 - Fix installation when manager directory name was changed
 
 ## [1.0.1-rc1] - 2012-01-11
+
 ### Changed
+
 - Refactor for MODx Revo 2.2
 
 ## [1.0.0-rc1] - 2011-11-16
+
 ### Changed
+
 - Packaged for Revo
 - Use ExtJS superboxselect type included in MODx Revolution
 
 ## [1.0.0-b1] - 2011-11-01
+
 ### Added
+
 - Initial Version