+++ core/components/superboxselect/docs/changelog.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.5] - 2022-05-17
+
+### Fixed
+
+- resourceTitleTpl and userTitleTpl for custom title in the SuperboxSelect row
+
 ## [3.0.4] - 2022-03-23
 
 ### Fixed